Updated controller and publish encoder info
Updated the PI in the Arduino code and added a digital filter to smooth the PWM signal sent to the motors in order to avoid an unstable system.

The ROS2 node now receives information from the Arduino with information about the encoders. This information is later published in newly made topics.